+++ lending/RELEASES.md
@@ -1169,3 +1169,13 @@ Based on:
 - [typescript v0.7.44] lending
 ### Releases
 - [NPM v0.7.44] https://www.npmjs.com/package/ryan-lending/v/0.7.44 - lending
+
+## 2025-01-09 00:58:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.465.1 (2.488.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.7.45] lending
+### Releases
+- [NPM v0.7.45] https://www.npmjs.com/package/ryan-lending/v/0.7.45 - lending